fix(lint): clear RUF046 + RUF059 from baseline ignore (Push 2 step 3)
Resolved 5 violations across 3 files. Removed 2 rule codes.

- RUF046 (3): `int(math.ceil(...))` -> `math.ceil(...)` (math.ceil
  already returns int in Python 3)
    - detections_3d.py:124,126
    - pc_preprocessing.py:334
- RUF059 (2): unused unpack -> _-prefix
    - graph_clustering.py:69  — `dists, idxs` -> `_dists, idxs`
    - graph_clustering.py:450 — `uniq, labels` -> `_uniq, labels`

Remaining temp-ignore (8 codes, ~138 violations):
  E501 ×81, F405 ×14, RUF003 ×13, E402 ×11, B905 ×10, RUF002 ×10,
  RUF001 ×4, F403 ×2.

+++ src/tls2dseg/detections_3d.py
@@ -121,9 +121,9 @@ def clean_pcd_instances_and_get_detections3d(
             if npts_i > min_npts:
                 expected_point_spacing = pcp_parameters["output_resolution"] * np.sqrt(3) * 1.1
                 clusterer_type = "hdbscan"  # 'dbscan', 'hdbscan'
-                min_cluster_size = int(math.ceil(0.25 * npts_i))
+                min_cluster_size = math.ceil(0.25 * npts_i)
                 if clusterer_type == "hdbscan":
-                    min_samples = int(math.ceil(0.005 * npts_i))
+                    min_samples = math.ceil(0.005 * npts_i)
                 elif clusterer_type == "dbscan":
                     min_samples = 5
